pyproject.toml
[project]
name = "mkwebsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "coverage>=7.13.5",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.6",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "redis>=7.2.1",
]
